requirements.txt
python-telegram-bot
python-dotenv
requests
cherrypy
jinja2
